refact(web): move gatsby config to its own folder
