+++ web/config-gatsby/index.js
@@ -1,0 +1,13 @@
+const siteMetadata = require('./siteMetadata')
+const filesystem = require('./filesystem')
+const style = require('./style')
+const seo = require('./seo')
+const transformers = require('./transformers')
+
+module.exports = {
+  siteMetadata,
+  filesystem,
+  style,
+  seo,
+  transformers,
+}
