--- web/site-metadata.json
@@ -1,9 +1,0 @@
-{
-  "siteUrl": "https://taverasmisael.com",
-  "title": "Misael Taveras",
-  "keywords": "taverasmisasel, Misael Taveras, Javascript, Frontend, React, Angular, Functional Programming, TDD, Programación Funcional, VueJS, CSS, SASS",
-  "description": "Blog personal de Misael Taveras. Hablemos de desarrollo frontend, Javascript, React, Angular, CSS y técnicas  de testing y arquitectura. De lo fundamental a lo más avanzado.",
-  "author": "Misael Taveras",
-  "socialHandler": "@taverasmisael",
-  "footerCopy": "Me encanta programar y compartir lo que he aprendido, asi que cree un website justo para eso"
-}
