+++ web/src/utils/social.js
@@ -1,2 +1,2 @@
-import { siteUrl } from '../../site-metadata.json'
+import { siteUrl } from '../../config-gatsby/siteMetadata'
 export const createCanonicalURL = url => encodeURI(`${siteUrl}${url}`)
